pyproject.toml
[tool.poetry]
name = "assignment-1"
version = "0.1.0"
description = "This is Querypie internship assignment"
authors = ["Scanf-s <calzone0404@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
